- Provide a shared bootstrap data module
The data is taken from the actual j1 schedule, but enriched with ID's for speaker, session and schedule.
The associations are kept through foreign key relationships.
A @Producer has been added to @Inject BootstrapData into other services that need the initial data to setup the demo.

- Leverage bootstrap data in session service

Currently the session service implements the bare bone read operation to retrieve session information. The implementation to remove or modify session data is still missing.