Demonstrate MP Health in vote microservice
Adds 3 Health check points on 3 DAO objects in the Vote microservice.
Added a link in the Vote UI that will take the user to
http://localhost:7070/health/ which will display a JSON response of the
health checks.  Eventually the UI could just call the health check and
parse the JSON to give a nice green/red indicator in the UI, but I don't
know enough about front end design to do that =)

Signed-off-by: Andrew Guibert <aguibert@us.ibm.com>